% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levels.R
\name{get_levels}
\alias{get_levels}
\title{Get municipalities levels}
\usage{
get_levels(
  start_period = NULL,
  end_period = NULL,
  escape_chars = NULL,
  hist_id = NULL,
  label_languages = c("", "de", "fr", "it", "en"),
  name_repair = "unique"
)
}
\arguments{
\item{start_period}{Start of requested period}

\item{end_period}{End of requested period}

\item{escape_chars}{Defines forbidden characters in the response and their
respective escape characters. For example, with the value "/( ).(_)",
all slashes in the response will be replaced with a space while periods are
replaced with underscore.}

\item{hist_id}{Determines whether BFSNR (snapshot) or HISTID
(geographic levels) will be delivered.}

\item{label_languages}{Include labels for the specified languages. Languages
are specified using the ISO 639-1 code (i.e., French ("fr"), German ("de"),
Italian ("it") and English ("en")) and are separated by a comma ",".}

\item{name_repair}{\code{\link[readr]{read_csv}} is used internally and
ensures column names are "unique" using the "name_repeir" argument.}
}
\value{
a data.frame/tibble
}
\description{
The geographical levels offer several classifications of municipalities
according to, for example, linguistic regions, agglomerations or even the
degree of urbanization.
}
\examples{
\donttest{
get_levels(
  start_period = "2024-01-01",
  end_period = "2024-08-01"
)
}

}
