% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutations.R
\name{get_mutations}
\alias{get_mutations}
\title{Get municipalities mutations}
\usage{
get_mutations(
  start_period = NULL,
  end_period = NULL,
  escape_chars = NULL,
  include_territory_exchange = NULL,
  name_repair = "unique"
)
}
\arguments{
\item{start_period}{Start of requested period}

\item{end_period}{End of requested period}

\item{escape_chars}{Defines forbidden characters in the response and their
respective escape characters. For example, with the value "/( ).(_)",
all slashes in the response will be replaced with a space while periods are
replaced with underscore.}

\item{include_territory_exchange}{Include/exclude records that only concern
territory changes.}

\item{name_repair}{\code{\link[readr]{read_csv}} is used internally and
ensures column names are "unique" using the "name_repeir" argument.}
}
\value{
a data.frame/tibble
}
\description{
Mutation list describes all changes related to municipalities that occurred
during the specified period.
}
\examples{
\donttest{
get_mutations(start_period = "2024-01-01", end_period = "2024-08-01")
}

}
