% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_muni_snapshots.R
\name{get_snapshots}
\alias{get_snapshots}
\title{Get municipalities snapshots}
\usage{
get_snapshots(
  start_period = Sys.Date(),
  end_period = Sys.Date(),
  escape_chars = NULL,
  hist_id = NULL,
  name_repair = "unique"
)
}
\arguments{
\item{start_period}{Start of requested period}

\item{end_period}{End of requested period}

\item{escape_chars}{Defines forbidden characters in the response and their
respective escape characters. For example, with the value "/( ).(_)",
all slashes in the response will be replaced with a space while periods are
replaced with underscore.}

\item{hist_id}{Determines whether BFSNR (snapshot) or
HISTID (geographic levels) will be delivered.}

\item{name_repair}{\code{\link[readr]{read_csv}} is used internally and
ensures column names are "unique" using the "name_repair" argument.}
}
\value{
a data.frame/tibble
}
\description{
a snapshot is the list of municipalities that exist for at least part of the
specified period (or the specified day when \code{start_period} = \code{end_period})
}
\examples{
\donttest{
get_snapshots(
  start_period = "2024-01-01",
  end_period = "2024-08-01"
)
}

}
